StudentData.cpp
This code organizes a data from individual students to put them in the right course. The struct for "Student" keeps track of their id number, name, and test score. The course struct keeps track of the course number, the data of the students in the course, and the number of students in the course. There are functions that allow the user to add students to a course, removes them, and print the data associated with each student.