Maze.cpp
This program asks the user to enter starting coordinates for two different racers. The user then enters a target coordinate for the racers. The last user entry is the moves for each racer. The program moves the racers and outputs if they reached the target and how many moves it took to reach the coordinates. This program primarily uses Classes.