skittlesbag.h
This is the .h for my skittles bag class that simulates a bag of skittles. The bag can be counted and poured out. Individual skittles can be eaten or added. The bag can be evened out to make each color skittle in the bag equal in amount. The count of each color of skittle in the bag can be determined. Different bags can be compared to see if they are exactly the same. A histogram graph of the bag of skittles can also be created from this class.