dynamicset.h
This is the .h file for my dynamic set class that is a data structure that utilizes memory to store and release integers in accordance to the structures needs. The structure essentially acts as an integer set but uses dynamic memory. The .h file was created as a class where we worked together to determine what functions were necessary for the class and what each function does. The .cpp file was written by me.