smoothieInheritance.cpp
This program is a derived class that inherits from a base class. Specifically, the base class is a general template for any type of drink. The smoothie class
inherits the drink class and then provides a template for any type of smoothie. This program allows for drink and/or smoothie objects to be created. The smoothie class only functions properly when it inherits the drink class.