wordSorter.cpp
This program reads words from a file into a data structure and outputs the alphabetized words line-by-line to a separate file without any duplicate words. It uses three different methods and also outputs the amount of time each method took into the console. The three methods mean using 2 different data structures to sort through the words. The first method uses a vector and checks for duplicates before inserting into the vector. The second method also uses a vector but uses the sort function in the vector class after every word has been inserted. The last method uses a set data structure, which means words won't be inserted if the word is already in the set. In order to use a certain method, the user must execute the compiled program like such: ./a.out "file name" "method number"; ex. ./a.out quickbrownfox.txt 2. There are two included txt files that the program has in place, but the program can be altered with different text files.