studentDatabase.cpp
This program uses the map data structure to create a password locked database of student records. The user will be prompted to enter a student ID number into the console and then provide the password of the student. If the password is correct, then the program will output the student's Name and GPA into the console. If the password is incorrect, then the user will be told it is an incorrect password. The file records.txt is required to run this program since it holds the database of information.