node.cpp
This node.cpp file contains the implementation of a node class along with recursive functions to do different tasks. The sequence function recursively finds the nth term in the sequence: a (subscript)n = a (subscript)n−2 + (n−1) for n≥2. The list_clear function recursively deletes the entire memory of a linked list. The reverse function recursively reverses an array of integers. The list_reverse function is not recursive, but it reverses a linked list.