OfficeHourSimulation.cpp
This program is a simulation of a college office hour. It runs through students asking the professor questions about certain topics and then sorts them. It also allows for the user to search which students asked certain topics and how many times each students came to the office hour. The information is outputted into three different txt files that hold different types of data. The simulations runs 100 times.